For static arrays, use sizeof(array), not magic numbers (cherry picked from commit 61d62f270d0135b07ea7a357d11dd66096596a0b)
Conflicts:
	src/common/image.h
	src/control/jobs/camera_jobs.c
	src/control/jobs/image_jobs.c
	src/gui/gtk.c

+++ src/common/image.h
@@ -135,7 +135,7 @@ void dt_image_film_roll_directory(const dt_image_t *img, char *pathname, int len
 /** returns the portion of the path used for the film roll name. */
 const char *dt_image_film_roll_name(const char *path);
 /** returns the film roll name, i.e. without the path. */
-void dt_image_film_roll(const dt_image_t *img, char *pathname, int len);
+void dt_image_film_roll(const dt_image_t *img, char *pathname, size_t len);
 /** appends version numbering for duplicated images. */
 void dt_image_path_append_version(int imgid, char *pathname, const int len);
 /** prints a one-line exif information string. */
